fix: remove the K parameter from mayo_algorithm
